requirements.txt
click
ecdsa
cryptography
dnspython
